added document helper and vectorstore helper which for RAG in AI
