+++ pom.xml
@@ -119,6 +119,13 @@
 			<version>8.0.1.Final</version>
 		</dependency>
 
+		<!-- PDF reader -->
+		<dependency>
+			<groupId>org.apache.pdfbox</groupId>
+			<artifactId>pdfbox</artifactId>
+			<version>2.0.27</version>
+		</dependency>
+
 		<!-- Junit -->
 		<dependency>
 			<groupId>org.junit.jupiter</groupId>
@@ -204,7 +211,7 @@
 			<dependency>
 				<groupId>org.springframework.ai</groupId>
 				<artifactId>spring-ai-bom</artifactId>
-				<version>1.0.0-SNAPSHOT</version>
+				<version>1.0.0-M6</version>
 				<type>pom</type>
 				<scope>import</scope>
 			</dependency>
