+++ src/main/java/com/ai/ai_bridge/helper/VectorStoreHelper.java
@@ -1,15 +1,10 @@
 package com.ai.ai_bridge.helper;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import com.ai.ai_bridge.model.ModelName;
 import org.springframework.ai.vectorstore.VectorStore;
-import org.springframework.core.io.Resource;
 
-public class VectorStoreHelper {
-    private static final Logger logger = LoggerFactory.getLogger(VectorStoreHelper.class);
+public interface VectorStoreHelper {
 
     //**
-    public void initializeVectorStore(VectorStore vectorStore, Resource file) {
-
-    }
+    VectorStore getVectorStore(ModelName modelName);
 }
