+++ src/main/resources/application.properties
@@ -12,6 +12,6 @@ spring.ai.openai.chat.options.temperature=0.7
 spring.ai.ollama.base-url= http://localhost:3000
 spring.ai.ollama.chat.enabled=true
 spring.ai.ollama.chat.options.model=llama3.2
-#spring.ai.ollama.embedding.enabled=true
-#spring.ai.ollama.embedding.options.model=nomic-embed-text
+spring.ai.ollama.embedding.enabled=true
+spring.ai.ollama.embedding.options.model=nomic-embed-text
